# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Review API",
        "description": "This is an open and free to use API for developers to test their client projects.\r\n        All API calls require you to pass in <br> your API Key with the querystring\r\n        <b>?apikey=########-####-####-####-############</b>, <br><br>Please\r\n        register for your own API Key today at https://reviews.simpleapi.dev/register ",
        "version": "1.0"
    },
    "paths": {
        "/api/reviews/{slug}/comments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create a comment on a review",
                "operationId": "comment.post",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of review to attach comment to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit comment field values",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "author_email": {
                                        "type": "string"
                                    },
                                    "author_slug": {
                                        "type": "string"
                                    },
                                    "meta": {
                                        "type": "object"
                                    },
                                    "datetime": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        },
                                        "author_email": {
                                            "type": "string"
                                        },
                                        "author_slug": {
                                            "type": "string"
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "datetime": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/{slug}/comments/{id}": {
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Update a comment on a review",
                "operationId": "comment.put",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of review to PUT comment on",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit comment field values",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "author_email": {
                                        "type": "string"
                                    },
                                    "author_slug": {
                                        "type": "string"
                                    },
                                    "meta": {
                                        "type": "object"
                                    },
                                    "datetime": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        },
                                        "author_email": {
                                            "type": "string"
                                        },
                                        "author_slug": {
                                            "type": "string"
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "datetime": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/{slug}/comment/{id}": {
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Deletes a specified comment on a review",
                "operationId": "comment.remove",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of review comment belongs to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a message"
                    }
                }
            }
        },
...
    }
}